requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.7
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
pydantic>=2.5.0
aiohttp>=3.9.0
lxml>=4.9.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0
fake-useragent>=1.4.0